Moved tests for reference/multitypereference
