+++ src/MooVC.Architecture.Tests/Ddd/ReferenceTests/TestableAggregate.cs
