+++ src/MooVC.Architecture.Tests/Ddd/ReferenceTests/TestableReference{T1,T2,T3,T4}.cs
