+++ src/MooVC.Architecture.Tests/Ddd/ReferenceTests/TestableReference{T1,T2}.cs
