+++ src/MooVC.Architecture.Tests/Ddd/ReferenceTests/WhenImplicitlyCastToReference.cs
